fix: show poll votes in post action
https://mask.atlassian.net/browse/FW-2247
https://mask.atlassian.net/browse/FW-2248
https://mask.atlassian.net/browse/FW-2206
