+++ src/components/Poll/PollCard.tsx
@@ -1,12 +1,9 @@
-import { Plural, t } from '@lingui/macro';
-import dayjs from 'dayjs';
 import { sumBy } from 'lodash-es';
 import { Fragment } from 'react';
 
 import { VoteButton } from '@/components/Poll/VoteButton.js';
 import { VoteResult } from '@/components/Poll/VoteResult.js';
 import { POLL_ACTION_ENABLED } from '@/constants/poll.js';
-import { getPollTimeLeft } from '@/helpers/getPollTimeLeft.js';
 import { isSameProfile } from '@/helpers/isSameProfile.js';
 import { useCurrentProfile } from '@/hooks/useCurrentProfile.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
@@ -31,13 +28,6 @@ export function PollCard({ post, frameUrl }: PollCardProps) {
         poll.options.some((option) => option.isVoted) ||
         isSameProfile(profile, post.author);
 
-    const timeLeft =
-        poll.votingStatus === 'closed' || (poll.endDatetime && dayjs(poll.endDatetime).isBefore(new Date()))
-            ? t`Final results`
-            : poll.endDatetime
-              ? getPollTimeLeft(poll.endDatetime)
-              : '';
-
     return (
         <div>
             {poll.options.map((option, index) => (
@@ -49,10 +39,6 @@ export function PollCard({ post, frameUrl }: PollCardProps) {
                     )}
                 </Fragment>
             ))}
-            <div className="mt-3 text-xs leading-6 text-lightSecond">
-                <Plural value={totalVotes} one={`${totalVotes} Vote`} other={`${totalVotes} Votes`} />
-                {timeLeft ? ` · ${timeLeft}` : ''}
-            </div>
         </div>
     );
 }
